--- src/main/java/eu/usrv/legacylootgames/auxiliary/cheatEvents/CE_PlayerDeath.java
@@ -1,12 +1,0 @@
-package eu.usrv.legacylootgames.auxiliary.cheatEvents;
-
-import net.minecraft.entity.player.EntityPlayer;
-
-public class CE_PlayerDeath implements ICheatEvent {
-
-    @Override
-    public void doEvent(EntityPlayer pPlayer) {
-        pPlayer.setHealth(0.0F);
-        pPlayer.setDead();
-    }
-}
