--- src/main/java/eu/usrv/legacylootgames/auxiliary/cheatEvents/ICheatEvent.java
@@ -1,10 +1,0 @@
-
-package eu.usrv.legacylootgames.auxiliary.cheatEvents;
-
-
-import net.minecraft.entity.player.EntityPlayer;
-
-
-public interface ICheatEvent {
-    void doEvent(EntityPlayer pPlayer);
-}
